Add audio support
for non-safari, at least: safari doesn't let you play audio that isn't
triggered by a user action like a click. In an ideal world they'd let
you request permission to do this, but alas. I use this stuff on my work
computer via chrome anyway so it's not the end of the world.